reafctor: userInfo (userInfo와 userDevice 분리, tracker 생성(load시 실행) )
